refactor: reorganize config module
